pyproject.toml
[project]
name = "finance-agent"
version = "0.1.0"
description = "AI-powered day trading agent using Alpaca Markets"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alpaca-py>=0.43",
    "anthropic>=0.45",
    "beautifulsoup4>=4.12",
    "earningscall>=1.4",
    "edgartools>=5.16",
    "feedparser>=6.0",
    "finnhub-python>=2.4",
    "httpx",
    "pydantic>=2.0",
]

[project.scripts]
finance-agent = "finance_agent.cli:main"

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "mypy",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["edgar.*", "edgar", "earningscall.*", "earningscall", "finnhub.*", "finnhub", "feedparser.*", "feedparser"]
ignore_missing_imports = true
follow_imports = "skip"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
